requirements.txt
alembic==1.4.0
astroid==2.3.3
autopep8==1.5
blinker==1.4
certifi==2019.11.28
Click==7.0
colorama==0.4.3
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
Flask-CKEditor==0.4.3
Flask-DebugToolbar==0.11.0
Flask-FlatPages==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-PageDown==0.2.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.0
lazy-object-proxy==1.4.3
Mako==1.1.1
Markdown==3.2.1
MarkupSafe==1.1.1
mccabe==0.6.1
pathtools==0.1.2
pipenv==2018.11.26
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.5.2
pylint==2.4.4
python-dateutil==2.8.1
python-dotenv==0.10.5
python-editor==1.0.4
python-markdown-math==0.6
PyYAML==5.3
six==1.14.0
SQLAlchemy==1.3.13
typed-ast==1.4.1
virtualenv==16.7.9
virtualenv-clone==0.5.3
watchdog==0.10.1
webencodings==0.5.1
Werkzeug==0.16.1
wrapt==1.11.2
WTForms==2.2.1
